requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-arm64
absl-py=1.4.0=pypi_0
aiohttp=3.8.4=pypi_0
aiosignal=1.3.1=pypi_0
anyio=3.6.2=pyhd8ed1ab_0
aom=3.5.0=h7ea286d_0
appnope=0.1.3=pyhd8ed1ab_0
argon2-cffi=21.3.0=pyhd8ed1ab_0
argon2-cffi-bindings=21.2.0=py310h8e9501a_3
asttokens=2.2.1=pyhd8ed1ab_0
astunparse=1.6.3=pypi_0
async-timeout=4.0.2=pypi_0
attrs=22.2.0=pyh71513ae_0
babel=2.11.0=pyhd8ed1ab_0
backcall=0.2.0=pyh9f0ad1d_0
backports=1.0=pyhd8ed1ab_3
backports.functools_lru_cache=1.6.4=pyhd8ed1ab_0
beautifulsoup4=4.11.2=pyha770c72_0
black=23.1.0=py310hbe9552e_0
bleach=6.0.0=pyhd8ed1ab_0
brotlipy=0.7.0=py310h8e9501a_1005
bzip2=1.0.8=h3422bc3_4
c-ares=1.18.1=h3422bc3_0
ca-certificates=2022.12.7=h4653dfc_0
cached-property=1.5.2=hd8ed1ab_1
cached_property=1.5.2=pyha770c72_1
cachetools=5.3.0=pypi_0
certifi=2022.12.7=pypi_0
cffi=1.15.1=py310h2399d43_3
charset-normalizer=3.0.1=pypi_0
click=8.1.3=unix_pyhd8ed1ab_2
comm=0.1.2=pyhd8ed1ab_0
contourpy=1.0.7=pypi_0
cryptography=39.0.0=py310hfc83b78_0
cycler=0.11.0=pypi_0
datasets=2.9.0=pypi_0
debugpy=1.6.6=py310h0f1eb42_0
decorator=5.1.1=pyhd8ed1ab_0
defusedxml=0.7.1=pyhd8ed1ab_0
dill=0.3.6=pypi_0
entrypoints=0.4=pyhd8ed1ab_0
executing=1.2.0=pyhd8ed1ab_0
expat=2.5.0=hb7217d7_0
ffmpeg=5.1.2=gpl_hf318d42_106
filelock=3.9.0=pypi_0
flatbuffers=1.12=pypi_0
flit-core=3.8.0=pyhd8ed1ab_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h82840c6_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.38.0=pypi_0
freetype=2.12.1=hd633e50_1
frozenlist=1.3.3=pypi_0
fsspec=2023.1.0=pypi_0
gast=0.4.0=pypi_0
gettext=0.21.1=h0186832_0
gmp=6.2.1=h9f76cd9_0
gmpy2=2.1.2=py310h2e6cad2_1
gnutls=3.7.8=h9f1a10d_0
google-auth=2.16.0=pypi_0
google-auth-oauthlib=0.4.6=pypi_0
google-pasta=0.2.0=pypi_0
grpcio=1.46.3=py310ha26ec5d_0
h5py=3.6.0=nompi_py310hb8bbf05_100
hdf5=1.12.1=nompi_hd9dbc9e_104
huggingface-hub=0.12.0=pypi_0
icu=70.1=h6b3803e_0
idna=3.4=pyhd8ed1ab_0
importlib-metadata=6.0.0=pyha770c72_0
importlib_metadata=6.0.0=hd8ed1ab_0
importlib_resources=5.10.2=pyhd8ed1ab_0
ipykernel=6.21.0=pyh736e0ef_0
ipython=8.9.0=pyhd1c38e8_0
ipython_genutils=0.2.0=py_1
ipywidgets=8.0.4=pyhd8ed1ab_0
jedi=0.18.2=pyhd8ed1ab_0
jinja2=3.1.2=pyhd8ed1ab_1
jpeg=9e=he4db4b2_2
json5=0.9.5=pyh9f0ad1d_0
jsonschema=4.17.3=pyhd8ed1ab_0
jupyter=1.0.0=py310hbe9552e_8
jupyter_client=8.0.2=pyhd8ed1ab_0
jupyter_console=6.4.4=pyhd8ed1ab_0
jupyter_core=5.2.0=py310hbe9552e_0
jupyter_events=0.6.3=pyhd8ed1ab_0
jupyter_server=2.2.0=pyhd8ed1ab_0
jupyter_server_terminals=0.4.4=pyhd8ed1ab_1
jupyterlab=3.5.3=pyhd8ed1ab_0
jupyterlab_pygments=0.2.2=pyhd8ed1ab_0
jupyterlab_server=2.19.0=pyhd8ed1ab_0
jupyterlab_widgets=3.0.5=pyhd8ed1ab_0
keras=2.9.0=pypi_0
keras-preprocessing=1.1.2=pypi_0
kiwisolver=1.4.4=pypi_0
krb5=1.20.1=h69eda48_0
lame=3.100=h1a8c8d9_1003
lcms2=2.14=h481adae_1
lerc=4.0.0=h9a09cb3_0
libblas=3.9.0=16_osxarm64_openblas
libcblas=3.9.0=16_osxarm64_openblas
libclang=15.0.6.1=pypi_0
libcurl=7.87.0=h9049daf_0
libcxx=14.0.6=h2692d47_0
libdeflate=1.17=h1a8c8d9_0
libedit=3.1.20191231=hc8eb9b7_2
libev=4.33=h642e427_1
libffi=3.4.2=h3422bc3_5
libgfortran=5.0.0=11_3_0_hd922786_27
libgfortran5=11.3.0=hdaf2cc0_27
libiconv=1.17=he4db4b2_0
libidn2=2.3.4=h1a8c8d9_0
libjpeg-turbo=2.1.4=h1a8c8d9_0
liblapack=3.9.0=16_osxarm64_openblas
libnghttp2=1.51.0=hae82a92_0
libopenblas=0.3.21=openmp_hc731615_3
libopus=1.3.1=h27ca646_1
libpng=1.6.39=h76d750c_0
libprotobuf=3.19.4=hccf11d3_0
libsodium=1.0.18=h27ca646_1
libsqlite=3.40.0=h76d750c_0
libssh2=1.10.0=h7a5bd25_3
libtasn1=4.19.0=h1a8c8d9_0
libtiff=4.5.0=h5dffbdd_2
libunistring=0.9.10=h3422bc3_0
libvpx=1.11.0=hc470f4d_3
libwebp-base=1.2.4=h57fd34a_0
libxcb=1.13=h9b22ae9_1004
libxml2=2.10.3=h87b0503_0
libzlib=1.2.13=h03a7124_4
llvm-openmp=15.0.7=h7cfbb63_0
markdown=3.4.1=pypi_0
markupsafe=2.1.2=py310h8e9501a_0
matplotlib=3.6.3=pypi_0
matplotlib-inline=0.1.6=pyhd8ed1ab_0
mistune=2.0.4=pyhd8ed1ab_0
mpc=1.3.1=h91ba8db_0
mpfr=4.1.0=h6d7a090_1
mpmath=1.2.1=pyhd8ed1ab_0
multidict=6.0.4=pypi_0
multiprocess=0.70.14=pypi_0
mypy_extensions=1.0.0=pyha770c72_0
nbclassic=0.4.8=pyhd8ed1ab_0
nbclient=0.7.2=pyhd8ed1ab_0
nbconvert=7.2.9=pyhd8ed1ab_0
nbconvert-core=7.2.9=pyhd8ed1ab_0
nbconvert-pandoc=7.2.9=pyhd8ed1ab_0
nbformat=5.7.3=pyhd8ed1ab_0
ncurses=6.3=h07bb92c_1
nest-asyncio=1.5.6=pyhd8ed1ab_0
nettle=3.8.1=h63371fa_1
networkx=3.0=pypi_0
notebook=6.5.2=pyha770c72_1
notebook-shim=0.2.2=pyhd8ed1ab_0
numpy=1.23.2=py310h127c7cf_0
oauthlib=3.2.2=pypi_0
openh264=2.3.1=hb7217d7_1
openjpeg=2.5.0=hbc2ba62_2
openssl=3.0.8=h03a7124_0
opt-einsum=3.3.0=pypi_0
p11-kit=0.24.1=h29577a5_0
packaging=23.0=pyhd8ed1ab_0
pandas=1.5.3=pypi_0
pandoc=2.19.2=hce30654_1
pandocfilters=1.5.0=pyhd8ed1ab_0
parso=0.8.3=pyhd8ed1ab_0
pathspec=0.11.0=pyhd8ed1ab_0
pexpect=4.8.0=pyh1a96a4e_2
pickleshare=0.7.5=py_1003
pillow=9.4.0=py310h5f2e7f5_0
pip=23.0=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pyhd8ed1ab_0
platformdirs=2.6.2=pyhd8ed1ab_0
prometheus_client=0.16.0=pyhd8ed1ab_0
prompt-toolkit=3.0.36=pyha770c72_0
prompt_toolkit=3.0.36=hd8ed1ab_0
protobuf=3.19.4=py310h1b49c16_0
psutil=5.9.4=py310h8e9501a_0
pthread-stubs=0.4=h27ca646_1001
ptyprocess=0.7.0=pyhd3deb0d_0
pure_eval=0.2.2=pyhd8ed1ab_0
pyarrow=11.0.0=pypi_0
pyasn1=0.4.8=pypi_0
pyasn1-modules=0.2.8=pypi_0
pycparser=2.21=pyhd8ed1ab_0
pygments=2.14.0=pyhd8ed1ab_0
pyopenssl=23.0.0=pyhd8ed1ab_0
pyparsing=3.0.9=pypi_0
pyrsistent=0.19.3=py310h8e9501a_0
pysocks=1.7.1=pyha2e5f31_6
python=3.10.8=h3ba56d0_0_cpython
python-dateutil=2.8.2=pyhd8ed1ab_0
python-dotenv=1.0.0=pypi_0
python-fastjsonschema=2.16.2=pyhd8ed1ab_0
python-json-logger=2.0.4=pyhd8ed1ab_0
python_abi=3.10=3_cp310
pytorch=2.0.0.dev20230202=py3.10_0
pytz=2022.7.1=pyhd8ed1ab_0
pyyaml=6.0=py310h8e9501a_5
pyzmq=25.0.0=py310hc407298_0
readline=8.1.2=h46ed386_0
regex=2022.10.31=pypi_0
requests=2.28.2=pyhd8ed1ab_0
requests-oauthlib=1.3.1=pypi_0
responses=0.18.0=pypi_0
rfc3339-validator=0.1.4=pyhd8ed1ab_0
rfc3986-validator=0.1.1=pyh9f0ad1d_0
rsa=4.9=pypi_0
send2trash=1.8.0=pyhd8ed1ab_0
setuptools=67.1.0=pyhd8ed1ab_0
six=1.15.0=pypi_0
sniffio=1.3.0=pyhd8ed1ab_0
soupsieve=2.3.2.post1=pyhd8ed1ab_0
stack_data=0.6.2=pyhd8ed1ab_0
svt-av1=1.4.1=h7ea286d_0
sympy=1.11.1=py310hbe9552e_2
tensorboard=2.9.1=pypi_0
tensorboard-data-server=0.6.1=pypi_0
tensorboard-plugin-wit=1.8.1=pypi_0
tensorflow-deps=2.10.0=0
tensorflow-estimator=2.9.0=pypi_0
tensorflow-macos=2.9.0=pypi_0
tensorflow-metal=0.5.0=pypi_0
termcolor=2.2.0=pypi_0
terminado=0.17.1=pyhd1c38e8_0
tinycss2=1.2.1=pyhd8ed1ab_0
tk=8.6.12=he1e0b03_0
tokenizers=0.13.2=pypi_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=2.0.0.dev20230202=py310_cpu
torchvision=0.15.0.dev20230202=py310_cpu
tornado=6.2=py310h8e9501a_1
tqdm=4.64.1=pypi_0
traitlets=5.9.0=pyhd8ed1ab_0
transformers=4.26.1=pypi_0
typing-extensions=4.4.0=hd8ed1ab_0
typing_extensions=4.4.0=pyha770c72_0
tzdata=2022g=h191b570_0
urllib3=1.26.14=pyhd8ed1ab_0
wcwidth=0.2.6=pyhd8ed1ab_0
webencodings=0.5.1=py_1
websocket-client=1.5.0=pyhd8ed1ab_0
werkzeug=2.2.2=pypi_0
wheel=0.38.4=pyhd8ed1ab_0
widgetsnbextension=4.0.5=pyhd8ed1ab_0
wrapt=1.14.1=pypi_0
x264=1!164.3095=h57fd34a_2
x265=3.5=hbc6ce65_3
xorg-libxau=1.0.9=h27ca646_0
xorg-libxdmcp=1.1.3=h27ca646_0
xxhash=3.2.0=pypi_0
xz=5.2.6=h57fd34a_0
yaml=0.2.5=h3422bc3_2
yarl=1.8.2=pypi_0
zeromq=4.3.4=hbdafb3b_1
zipp=3.12.0=pyhd8ed1ab_0
zlib=1.2.13=h03a7124_4
zstd=1.5.2=hf913c23_6
